% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{biodiv_effect_K10}
\alias{biodiv_effect_K10}
\title{An STM regression object for a topic model with 10 topics}
\format{
An STM regression object
}
\usage{
biodiv_effect_K10
}
\description{
This package data set provides the results of applying the
\code{\link[stm:estimateEffect]{stm::estimateEffect()}} function to the included \code{biodiv_model_K10} topic
model object.
}
\keyword{datasets}
