% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{biodiv_model_K10}
\alias{biodiv_model_K10}
\title{An STM topic model with 10 topics}
\format{
An STM topic model object
}
\usage{
biodiv_model_K10
}
\description{
This package data set provides a sample STM topic model with K = 10 topics
(see \code{\link[stm:stm]{stm::stm()}}).
The model has been fit to bioRxiv and medRxiv preprints from 2017 to 2021 that
contain the term 'biodiversity' in the title or abstract. The model
has been fit with the respective publication status and publication year as
topic prevalence covariates. The included \code{biodiv_effect_K10} dataset captures
the regression parameters of these document covariates.
}
\keyword{datasets}
