% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{covid_effect_K20}
\alias{covid_effect_K20}
\title{An STM regression object for a topic model with 20 topics}
\format{
An STM regression object
}
\usage{
covid_effect_K20
}
\description{
This package data set provides the results of applying the
\code{\link[stm:estimateEffect]{stm::estimateEffect()}} function to the included \code{covid_effect_K20} topic
model object.
}
\keyword{datasets}
