% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{covid_model_K20}
\alias{covid_model_K20}
\title{An STM topic model with 20 topics}
\format{
An STM topic model object
}
\usage{
covid_model_K20
}
\description{
This package data set provides a sample STM topic model with K = 20 topics
(see \code{\link[stm:stm]{stm::stm()}}).
The model has been fit to bioRxiv and medRxiv preprints from 2020 to 2023 that
contain the terms 'sars-cov' or 'covid' in the title or abstract. The model
has been fit with the respective preprint server and publication year as
topic prevalence covariates. The included \code{covid_effect_K20} dataset captures
the regression parameters of these document covariates.
}
\keyword{datasets}
