% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcquantcourse.R
\docType{package}
\name{srcquantcourse-package}
\alias{srcquantcourse}
\alias{srcquantcourse-package}
\title{srcquantcourse: Scripts and Exercises for the SRC Quantitative Methods PhD Course}
\description{
This package collects scripts, data and exercises notes for the 2024 SRC
Quantitative Methods PhD course.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://sdaume.github.io/srcquantcourse}
  \item \url{https://github.com/sdaume/srcquantcourse}
  \item Report bugs at \url{https://github.com/sdaume/srcquantcourse/issues}
}

}
\author{
\strong{Maintainer}: Stefan Daume \email{stefan.daume@su.se} (\href{https://orcid.org/0000-0002-5879-0290}{ORCID})

}
