% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcquantcourse.R
\docType{package}
\name{srcquantcourse}
\alias{srcquantcourse}
\title{Scripts and exercises for the SRC Quantitative Methods PhD Course}
\description{
This package collects scripts, data and exercises notes for the 2024 SRC
Quantitative Methods PhD course.
}
