% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{topicmodel_evaluations}
\alias{topicmodel_evaluations}
\title{Evaluation results for topic models with different K}
\format{
A dataframe with four variables
}
\usage{
topicmodel_evaluations
}
\description{
This is a convenience dataset to include the results of evaluating topic
models with different topic numbers (\emph{K}) for thematic subsets of the
bioRxiv and medRxiv preprints. Scripts to run the actual evaluations are
included in the package repo.
}
\details{
The dataset specifies the preprint subset, used K, type of evaluation metric
and the value of that metric.
}
\keyword{datasets}
